@@ -1,8 +1,8 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 26
-    buildToolsVersion "26.0.0"
+    compileSdkVersion 29
+    buildToolsVersion '28.0.3'
     defaultConfig {
         applicationId "com.youcoupon.john_li.transportationapp"
         minSdkVersion 21
@@ -11,6 +11,9 @@ android {
         versionName "2.0.14"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
         multiDexEnabled true
+        ndk {
+            abiFilters "armeabi" //, "x86", "armeabi"  按需设置
+        }
     }
     buildTypes {
         release {
@@ -18,22 +21,27 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+    sourceSets {
+        main.jniLibs.srcDirs = ['libs']
+        main{
+            jniLibs.srcDirs = ['libs']
+        }
+    }
     productFlavors {
     }
 }
 
 dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+    androidTestImplementation('com.android.support.test.espresso:espresso-core:2.2.2', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
-    compile 'com.android.support:appcompat-v7:26.+'
-    compile 'com.android.support:design:26.0.0-alpha1'
-    compile 'com.android.support.constraint:constraint-layout:1.0.2'
-    testCompile 'junit:junit:4.12'
-    compile files('libs/core-2.3.0.jar')
-    compile 'org.xutils:xutils:3.5.0'
-    compile 'com.google.code.gson:gson:2.8.1'
-    compile 'org.greenrobot:eventbus:3.0.0'
-    compile 'com.google.code.findbugs:jsr305:2.0.1'
+    testImplementation 'junit:junit:4.13'
+    implementation files('libs/core-2.3.0.jar')
+    implementation 'org.xutils:xutils:3.5.0'
+    implementation 'com.google.code.gson:gson:2.8.6'
+    implementation 'org.greenrobot:eventbus:3.0.0'
+    implementation 'com.google.code.findbugs:jsr305:3.0.2'
+    implementation 'androidx.fragment:fragment:1.2.4'
+    implementation 'com.google.android.material:material:1.0.0'
 }
