@@ -1,7 +1,8 @@
 package com.youcoupon.john_li.transportationapp.TMSActivity;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+
+import androidx.annotation.Nullable;
 
 import com.youcoupon.john_li.transportationapp.R;
 
