@@ -1,10 +1,11 @@
 package com.youcoupon.john_li.transportationapp.TMSActivity;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
 import android.view.View;
 import android.widget.ListView;
 import android.widget.Toast;
+
+import androidx.annotation.Nullable;
 
 import com.google.gson.Gson;
 import com.youcoupon.john_li.transportationapp.R;
