@@ -8,16 +8,18 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.annotation.RequiresApi;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.EditText;
+import android.widget.LinearLayout;
 import android.widget.ListView;
 import android.widget.RadioButton;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import androidx.annotation.Nullable;
+import androidx.annotation.RequiresApi;
 
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
@@ -25,11 +27,14 @@ import com.youcoupon.john_li.transportationapp.R;
 import com.youcoupon.john_li.transportationapp.TMSAdapter.DeliverGoodsAdapter;
 import com.youcoupon.john_li.transportationapp.TMSDBInfo.CustomerInfo;
 import com.youcoupon.john_li.transportationapp.TMSDBInfo.InvoiceInfo;
+import com.youcoupon.john_li.transportationapp.TMSDBInfo.MaterialCorrespondenceInfo;
 import com.youcoupon.john_li.transportationapp.TMSDBInfo.MaterialNumberInfo;
 import com.youcoupon.john_li.transportationapp.TMSDBInfo.SubmitInvoiceInfo;
 import com.youcoupon.john_li.transportationapp.TMSModel.Barcodemode;
 import com.youcoupon.john_li.transportationapp.TMSModel.CommonModel;
 import com.youcoupon.john_li.transportationapp.TMSModel.DeliverInvoiceModel;
+import com.youcoupon.john_li.transportationapp.TMSModel.InvoiceViewModel;
+import com.youcoupon.john_li.transportationapp.TMSModel.MaterialCorrespondenceModel;
 import com.youcoupon.john_li.transportationapp.TMSModel.PostInvoiceModel;
 import com.youcoupon.john_li.transportationapp.TMSModel.UserModel;
 import com.youcoupon.john_li.transportationapp.TMSUtils.ScanHelper;
@@ -38,6 +43,7 @@ import com.youcoupon.john_li.transportationapp.TMSUtils.TMSApplication;
 import com.youcoupon.john_li.transportationapp.TMSUtils.TMSCommonUtils;
 import com.youcoupon.john_li.transportationapp.TMSUtils.TMSConfigor;
 import com.youcoupon.john_li.transportationapp.TMSUtils.TMSShareInfo;
+import com.youcoupon.john_li.transportationapp.TMSView.NoScrollListView;
 import com.youcoupon.john_li.transportationapp.TMSView.TMSHeadView;
 
 import org.json.JSONException;
@@ -53,6 +59,7 @@ import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
@@ -62,8 +69,9 @@ import java.util.Map;
 
 public class DeliverGoodsActivity extends BaseActivity implements View.OnClickListener{
     private TMSHeadView headView;
-    private TextView delivergoods_customer_no,delivergoods_customer_name, delivergoods_customer_address, delivergoods_tel, circleClockInTv, invoiceTv;
-    private ListView mLv;
+    private TextView delivergoods_customer_no,delivergoods_customer_name, delivergoods_customer_address, delivergoods_tel, scanInvoiceTv, circleClockInTv, invoiceTv;
+    private LinearLayout invoiceLL;
+    private NoScrollListView mLv;
     private ProgressDialog dialog;
 
     // 物料订单的数据库类
@@ -92,8 +100,10 @@ public class DeliverGoodsActivity extends BaseActivity implements View.OnClickLi
         delivergoods_customer_name = findViewById(R.id.delivergoods_customer_name);
         delivergoods_customer_address = findViewById(R.id.delivergoods_customer_address);
         delivergoods_tel = findViewById(R.id.delivergoods_tel);
+        scanInvoiceTv = findViewById(R.id.delivergoods_invoiceno);
         circleClockInTv = findViewById(R.id.shortcut_delivergoods_circle_clockin);
         invoiceTv = findViewById(R.id.shortcut_delivergoods_invoice);
+        invoiceLL = findViewById(R.id.delivergoods_invoiceno_ll);
     }
 
     @Override
@@ -104,7 +114,7 @@ public class DeliverGoodsActivity extends BaseActivity implements View.OnClickLi
 
     @Override
     public void initData() {
-        headView.setTitle("送貨");
+        headView.setTitle("物料回收");
         headView.setLeft(this);
 
         mSubmitInvoiceInfo = new SubmitInvoiceInfo();
@@ -139,13 +149,17 @@ public class DeliverGoodsActivity extends BaseActivity implements View.OnClickLi
     private void getData() {
         try {
             List<MaterialNumberInfo> all = TMSApplication.db.selector(MaterialNumberInfo.class).findAll();
-            for(MaterialNumberInfo model : all){
-                DeliverInvoiceModel deliverInvoiceModel = new DeliverInvoiceModel();
-                deliverInvoiceModel.setMaterialId(model.getMaterialId());
-                deliverInvoiceModel.setMaterialName(model.getMaterialName());
-                deliverInvoiceModel.setSendOutNum(0);
-                deliverInvoiceModel.setRecycleNum(0);
-                mDeliverInvoiceModelList.add(deliverInvoiceModel);
+            if(all != null) {
+                for(MaterialNumberInfo model : all){
+                    DeliverInvoiceModel deliverInvoiceModel = new DeliverInvoiceModel();
+                    deliverInvoiceModel.setMaterialId(model.getMaterialID());
+                    deliverInvoiceModel.setMaterialName(model.getNameChinese());
+                    deliverInvoiceModel.setSendOutNum(0);
+                    deliverInvoiceModel.setRecycleNum(0);
+                    mDeliverInvoiceModelList.add(deliverInvoiceModel);
+                }
+            } else {
+                Toast.makeText(this, "暫無物料信息，請檢查並更新！", Toast.LENGTH_SHORT).show();
             }
         } catch (DbException e) {
             e.printStackTrace();
@@ -160,6 +174,12 @@ public class DeliverGoodsActivity extends BaseActivity implements View.OnClickLi
                 finish();
                 break;
             case R.id.head_right_tv:
+                dialog = new ProgressDialog(this);
+                dialog.setTitle("提示");
+                dialog.setMessage("正在提交發票......");
+                dialog.setCancelable(false);
+                dialog.show();
+
                 boolean b = false;
                 for(DeliverInvoiceModel model : mDeliverInvoiceModelList) {
                     if (model.getSendOutNum() > 0 || model.getRecycleNum() > 0) {
@@ -182,9 +202,11 @@ public class DeliverGoodsActivity extends BaseActivity implements View.OnClickLi
                         checkInvoiceType(TMSShareInfo.IMEI + time);
                     } catch (DbException e) {
                         e.printStackTrace();
+                        dialog.dismiss();
                         Toast.makeText(DeliverGoodsActivity.this, "訂單存儲失敗，請重試！", Toast.LENGTH_SHORT).show();
                     }
                 } else {
+                    dialog.dismiss();
                     Toast.makeText(DeliverGoodsActivity.this, "暫無物料需要提交！", Toast.LENGTH_SHORT).show();
                 }
                 break;
@@ -224,8 +246,8 @@ public class DeliverGoodsActivity extends BaseActivity implements View.OnClickLi
             } catch (DbException e) {
                 e.printStackTrace();
             }
-            int d = mDeliverInvoiceModelList.get(i).getSendOutNum();
-            int r = mDeliverInvoiceModelList.get(i).getRecycleNum();
+            int sendOutNum = mDeliverInvoiceModelList.get(i).getSendOutNum();
+            int recycleNum = mDeliverInvoiceModelList.get(i).getRecycleNum();
             // 当数量大于0时记录到记录物料总数的表中(MaterialNumberInfo)
             if (mDeliverInvoiceModelList.get(i).getSendOutNum() > 0) {
                 try {
@@ -233,7 +255,6 @@ public class DeliverGoodsActivity extends BaseActivity implements View.OnClickLi
                     b.and("material_number_id","=", mDeliverInvoiceModelList.get(i).getMaterialId()); //构造修改的条件
                     KeyValue name = new KeyValue("material_deposite_num", depositNum + mDeliverInvoiceModelList.get(i).getSendOutNum());
                     TMSApplication.db.update(MaterialNumberInfo.class,b,name);
-                    totalDepositNum = depositNum + mDeliverInvoiceModelList.get(i).getSendOutNum();
                 } catch (DbException e) {
                     e.printStackTrace();
                 }
@@ -245,18 +266,27 @@ public class DeliverGoodsActivity extends BaseActivity implements View.OnClickLi
                     b.and("material_number_id","=", mDeliverInvoiceModelList.get(i).getMaterialId()); //构造修改的条件
                     KeyValue name = new KeyValue("material_refund_num", refundNum + mDeliverInvoiceModelList.get(i).getRecycleNum());
                     TMSApplication.db.update(MaterialNumberInfo.class,b,name);
-                    totalRefundNum = refundNum + mDeliverInvoiceModelList.get(i).getRecycleNum();
                 } catch (DbException e) {
                     e.printStackTrace();
                 }
             }
+
+            totalDepositNum += depositNum + mDeliverInvoiceModelList.get(i).getSendOutNum();
+            totalRefundNum += refundNum + mDeliverInvoiceModelList.get(i).getRecycleNum();
+
             /*depositNum += mDeliverInvoiceModelList.get(i).getSendOutNum();
             refundNum += mDeliverInvoiceModelList.get(i).getRecycleNum();*/
         }
 
+        // 将物料回收的资料提交到服务器
         if (totalRefundNum > 0) {
             invoiceResult = invoiceResult + 1;
             submitTimes ++;
+            try {
+                Thread.sleep(1000);
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
             // 提交物料回收发票
             callNetSubmitInvoice(totalDepositNum, totalRefundNum, 1, reference);
         } else {
@@ -291,11 +321,6 @@ public class DeliverGoodsActivity extends BaseActivity implements View.OnClickLi
      * @param reference
      */
     private void callNetSubmitInvoice(final int depositNum, final int refundNum, final int type, final String reference) {
-        dialog = new ProgressDialog(this);
-        dialog.setTitle("提示");
-        dialog.setMessage("正在提交發票......");
-        dialog.setCancelable(false);
-        dialog.show();
         Map<String, String> paramsMap = new HashMap<>();
         paramsMap.put("corp", TMSCommonUtils.getUserFor40(this).getCorp());
         paramsMap.put("userid", TMSCommonUtils.getUserFor40(this).getID());
@@ -314,16 +339,30 @@ public class DeliverGoodsActivity extends BaseActivity implements View.OnClickLi
             if (type == 0) {
                 header.setReference(submitInvoiceInfo.getRefrence());
             } else {
-                // 当为发票类型为回收物料时，判断送出物料是否为0，不为0则以送出物料为主单不修改发票号码(即不处理订单号码以送出单的编码为发票编码)，当送出物料数量为0时以回收物料为主单直接修以回收物料发票编码为发票编码)
+                // 当为发票类型为回收物料时，判断送出物料是否为0，不为0则以送出物料为主單
+                // 主单不修改发票号码(即不处理订单号码以送出单的编码为发票编码)，当送出物料数量为0时以回收物料为主单直接修以回收物料发票编码为发票编码)
                 if (depositNum == 0) {
                     header.setReference(submitInvoiceInfo.getRefrence());
                 } else {
-                    String time = new SimpleDateFormat("yyyy-MM-dd-HH:mm:ss").format(new Date()).replace("-","");
+                    String sunReference = TMSShareInfo.IMEI + new SimpleDateFormat("yyyy-MM-dd-HH:mm:ss").format(new Date()).replace("-", "").replace(":", "").replace(" ", "");
+                    mSubmitInvoiceInfo.setSunRefrence(sunReference);
+                    try {
+                        WhereBuilder b = WhereBuilder.b();
+                        b.and("refrence", "=", reference);//条件
+                        KeyValue keyValue = new KeyValue("sun_refrence", sunReference);
+                        TMSApplication.db.update(SubmitInvoiceInfo.class, b, keyValue);
+                    } catch (Exception e) {
+                        Toast.makeText(this, "666", Toast.LENGTH_SHORT).show();
+                    }
+                    header.setReference(sunReference);
+
+                    /*String time = new SimpleDateFormat("yyyy-MM-dd-HH:mm:ss").format(new Date()).replace("-","");
                     time = time.replace(":","");
                     time = time.replace(" ","");
-                    header.setReference(TMSShareInfo.IMEI + time);
+                    header.setReference(TMSShareInfo.IMEI + time);*/
                 }
             }
+
             header.setSalesmanID(TMSCommonUtils.getUserFor40(this).getSalesmanID());
             postInvoiceModel.setHeader(header);
             // 发票表体
@@ -358,15 +397,10 @@ public class DeliverGoodsActivity extends BaseActivity implements View.OnClickLi
 
                     invoiceResult = invoiceResult - 1;
                     if (type == 0) {
-                        // 修改发票号码
                         try {
+                            // 修改发票号码
                             TMSApplication.db.update(SubmitInvoiceInfo.class, WhereBuilder.b().and("refrence","=", reference),new KeyValue("invoice_no", orderNo));
-                        } catch (DbException e) {
-                            e.printStackTrace();
-                        }
-
-                        // 修改发票送出物料状态
-                        try {
+                            // 修改发票送出物料状态
                             TMSApplication.db.update(SubmitInvoiceInfo.class, WhereBuilder.b().and("refrence","=",reference),new KeyValue("depositStatus", 1));
                         } catch (DbException e) {
                             e.printStackTrace();
@@ -376,6 +410,12 @@ public class DeliverGoodsActivity extends BaseActivity implements View.OnClickLi
                         if (depositNum == 0) {
                             try {
                                 TMSApplication.db.update(SubmitInvoiceInfo.class, WhereBuilder.b().and("refrence","=",reference),new KeyValue("invoice_no", orderNo));
+                            } catch (DbException e) {
+                                e.printStackTrace();
+                            }
+                        } else {
+                            try {
+                                TMSApplication.db.update(SubmitInvoiceInfo.class, WhereBuilder.b().and("refrence","=",reference),new KeyValue("sun_invoice_no", orderNo));
                             } catch (DbException e) {
                                 e.printStackTrace();
                             }
@@ -474,6 +514,7 @@ public class DeliverGoodsActivity extends BaseActivity implements View.OnClickLi
                     Barcodemode code = new Barcodemode();
                     code.setBarcode(str);
                     code.setNumber("");
+                    invoiceLL.setVisibility(View.GONE);
                     IshavaInvoiceCode(code);
                 }
             }
@@ -496,6 +537,8 @@ public class DeliverGoodsActivity extends BaseActivity implements View.OnClickLi
             }
 
             if (invoiceInfo != null) {
+                invoiceLL.setVisibility(View.VISIBLE);
+                scanInvoiceTv.setText(invoiceInfo.getInvoiceNo());
                 CustomerInfo customerInfo = null;
                 List<CustomerInfo> customerInfoList = TMSApplication.db.selector(CustomerInfo.class).where("customer_id","=",invoiceInfo.getCustomerID()).findAll();
                 for(CustomerInfo info : customerInfoList){
@@ -509,6 +552,8 @@ public class DeliverGoodsActivity extends BaseActivity implements View.OnClickLi
                     delivergoods_customer_name.setText(customerInfo.getCustomerName());
                     delivergoods_customer_address.setText(customerInfo.getCustomerAddress());
                     delivergoods_tel.setText(customerInfo.getTelephone());
+                    // 填充發票中對應的物料
+                    fillInvoiceMaterial(invoiceInfo);
                     headView.setRightText("提交", this);
                     mLv.setOnItemClickListener(new AdapterView.OnItemClickListener() {
                         @Override
@@ -566,13 +611,53 @@ public class DeliverGoodsActivity extends BaseActivity implements View.OnClickLi
                     IshavaCustomerCode(code);
                 }
             } else {
-                Toast.makeText(DeliverGoodsActivity.this, "未找到該該發票！", Toast.LENGTH_SHORT).show();
+                //Toast.makeText(DeliverGoodsActivity.this, "未找到該該發票！", Toast.LENGTH_SHORT).show();
                 IshavaCustomerCode(code);
             }
         } catch (DbException e) {
             e.printStackTrace();
             Toast.makeText(DeliverGoodsActivity.this, "查找發票錯誤！", Toast.LENGTH_SHORT).show();
             IshavaCustomerCode(code);
+        }
+    }
+
+    // 填充發票貨品對應物料
+    private void fillInvoiceMaterial(InvoiceInfo invoiceInfo) {
+        // 發票Model轉ViewModel
+        List<InvoiceViewModel.InvoiceLine> lineList = new Gson().fromJson(invoiceInfo.getLines(), new TypeToken<List<InvoiceViewModel.InvoiceLine>>() {}.getType());
+        try {
+            List<MaterialCorrespondenceInfo> materialCorrespondenceList = TMSApplication.db.selector(MaterialCorrespondenceInfo.class).findAll();
+            List<MaterialCorrespondenceModel> materialCorrespondenceModelList = new ArrayList<>();
+            for (MaterialCorrespondenceInfo info : materialCorrespondenceList) {
+                MaterialCorrespondenceModel model = new MaterialCorrespondenceModel();
+                model.setMerchandiseID(info.getMerchandiseID().trim());
+                List<MaterialCorrespondenceModel.CorrespondingMaterial> materialList = new Gson().fromJson(info.getMaterialListJson(), new TypeToken<List<MaterialCorrespondenceModel.CorrespondingMaterial>>() {}.getType());
+                for (MaterialCorrespondenceModel.CorrespondingMaterial correspondingMaterial: materialList) {
+                    correspondingMaterial.setMaterialID(correspondingMaterial.getMaterialID().trim());
+                }
+                model.setMaterial(materialList);
+                materialCorrespondenceModelList.add(model);
+            }
+
+            for (InvoiceViewModel.InvoiceLine line: lineList){
+                for (MaterialCorrespondenceModel materialCorrespondenceModel : materialCorrespondenceModelList) {
+                    if (line.getMerchandiseID().equals(materialCorrespondenceModel.getMerchandiseID())) {
+
+                        for (DeliverInvoiceModel model : mDeliverInvoiceModelList){
+                            for (MaterialCorrespondenceModel.CorrespondingMaterial correspondingMaterialModel: materialCorrespondenceModel.getMaterial()) {
+                                if (model.getMaterialId().equals(correspondingMaterialModel.getMaterialID())) {
+                                    model.setSendOutNum(model.getSendOutNum() + (line.getQuantity() * correspondingMaterialModel.getQuantity()));
+                                    model.setRecycleNum(model.getRecycleNum() + (line.getQuantity() * correspondingMaterialModel.getQuantity()));
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+
+            mDeliverGoodsAdapter.notifyDataSetChanged();
+        } catch (DbException e) {
+            e.printStackTrace();
         }
     }
 
@@ -615,6 +700,7 @@ public class DeliverGoodsActivity extends BaseActivity implements View.OnClickLi
                         TextView cancel = view1.findViewById(R.id.dialog_dg_cancel_tv);
                         final RadioButton rb1 = view1.findViewById(R.id.dialog_dg_send_out_rb);
                         final RadioButton rb2 = view1.findViewById(R.id.dialog_dg_recycle_rb);
+                        final RadioButton rb3 = view1.findViewById(R.id.dialog_dg_send_recycle_rb);
 
                         final Dialog d = dialog1.create();
                         name.setText(mDeliverInvoiceModelList.get(position).getMaterialName());
@@ -623,10 +709,13 @@ public class DeliverGoodsActivity extends BaseActivity implements View.OnClickLi
                             public void onClick(View v) {
                                 try {
                                     if (!numEt.getText().equals("")) {
-                                        if ((rb1.isChecked() || rb2.isChecked() ) && !(rb1.isChecked() && rb2.isChecked())) {
+                                        if ((rb1.isChecked() || rb2.isChecked() || rb3.isChecked())) { //&& !(rb1.isChecked() && rb2.isChecked())
                                             if (rb1.isChecked()) {
                                                 mDeliverInvoiceModelList.get(position).setSendOutNum(Integer.parseInt(numEt.getText().toString()));
                                             } else if (rb2.isChecked()) {
+                                                mDeliverInvoiceModelList.get(position).setRecycleNum(Integer.parseInt(numEt.getText().toString()));
+                                            } else if (rb3.isChecked()) {
+                                                mDeliverInvoiceModelList.get(position).setSendOutNum(Integer.parseInt(numEt.getText().toString()));
                                                 mDeliverInvoiceModelList.get(position).setRecycleNum(Integer.parseInt(numEt.getText().toString()));
                                             }
                                             mDeliverGoodsAdapter.notifyDataSetChanged();
