@@ -209,7 +209,8 @@ public final class BarcodeUtil {
 		Canvas cv = new Canvas(newBitmap);
 		cv.drawBitmap(first, marginW, 0, null);
 		cv.drawBitmap(second, fromPoint.x, fromPoint.y, null);
-		cv.save(Canvas.ALL_SAVE_FLAG);
+		//cv.save(Canvas.ALL_SAVE_FLAG);
+		cv.save();
 		cv.restore();
 		return newBitmap;
 	}
